requirements.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
certifi==2021.5.30
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.4.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.3
pipenv==2022.1.8
platformdirs==2.4.1
protobuf==3.17.3
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
virtualenv-clone==0.5.4
wrapt==1.13.3
yapf==0.32.0
